requirements.txt
Flask==1.0.2
scikit-learn>=0.19
numpy>=1.14
scipy>=0.15.1

panda>=0.1.5